add Zimbra repository and packages on ubuntu
This mainly affects openldap, which requires a bit of permission
jiggling to work properly. I've extracted this logic to a function,
since we now need to install/create multiple directories with the
appropriate permissions, and applied it to creating '/opt/zimbra' as
well.

While at it, I relaxed the permissions a bit from world-writable to
writable by the 'sudo' group. Normally, developers will either just
use 'vagrant' user or create their own. If they do the latter, they'll
surely want that user to have 'sudo' privileges.