don't apply memory limit to LXC containers
Given that most of Zimbra is in Java, and Java is both extremely fond
of memory and extremely slow when it runs out, we shouldn't limit it
unless we really need to. When sharing a single kernel, individual
containers won't take up any more memory than they have to with a
higher limit, so the only effect of limiting memory is that it makes
everything slower.

At the very least, we should apply a higher limit; 2GB is not
sufficient for building Zimbra.