autodetect JAVA_HOME based on the javac in the PATH
A properly set JAVA_HOME is required for building Zimbra.  Without
this change one has to determine the proper path to the JDK supplied
by the OS everytime one does a vagrant ssh.

This sets the JAVA_HOME to the value corresponding to the javac found
in the PATH.  If the Zimbra javac should be used, the PATH should
be prepended with /opt/zimbra/common/bin to have it picked up.